Bugfix in calculating leading whitespace
The calculation for leading whitespace failed when the file contained an empty
line (because an empty line might now have enough whitespace). So, now we
explicitly check for blank lines.

A minor optimization for checking of leading white space. If the current value
is 0, break the loop. This might speed things up for big files.

Another optimization: Instead of translating the entire line and thenremoving
the leading white space, we now only truncate the line before translating.