Experimental feature: raw mode
Support for raw mode

\startrawvimtyping
....
\stoprawvimtyping

\inlinerawvimtyping{....}

This is useful if an external program is generating ConTeXt code and wants to
highlight using the vim module. Hopefully kate-highlighting will provide a
ConTeXt mode in the future.

We can also think of allowing the user to run 2context on big files and then
just input the parsed file.