Use lua to locate 2context file in MkIV
MkIV now uses job.files.run to check if a file has changed, instead of mtxrun
--ifchanged. This considerably speeds up the compilation since an external
program is not called at each run. However the drawback of this approach is that
we can no longer use `kpse:2context.vim` to get the location of `2context.vim`
in the filesystem. So we now rely on `resolvers.resolve`.