Make sure that line numbering when lines are split
The default line numbering mechanism numbers all the lines in the output. For
source code, a line can be split into multiple lines because of its length. In
that case, we only want the first line to be numbered.

ConTeXt already has this mechanism for \starttyping and \stoptyping. So, we just
hook into that mechanism.