Only define a new counter if it does not exist.
When using pandoc to generate context file, it is useful to redefine
the `typing` environment as a vimtyping environment using

    \definevimtyping[typing][...]

    markdown code

    \definevimtyping[typing][...]

    more markdown code

Currently, \definevimtyping effectively resets the counter, the first temp file
is overwritten by the second, and we only get one code listing in the output.
The current change fixes this bug.