Cleaner way to write output to a new directory
Previous code was using (pseudocode):

```
\savebuffer[list-of-buffers][temp-file]
\startluacode
  os.copy(temp-file, actual-file)
  os.remove(temp-file)
\stopluacode
```

The reason for this convoluted code was that `\savebuffer[...]` appends
a prefix `\jobname-`, which does not work when we want to write to a
directory.

Based on a [question] on TeX.SX, I realized that I could avoid this roundabout
route by using the key-value interface of `\savebuffer`, in which one can set
`prefix=no`.

[question]: https://tex.stackexchange.com/questions/372071/how-to-save-buffer-in-subfolder