\ctxcommand(getreadfilename) --> \clf_getreadfilename
There was a bug. Bugreport on ntg-context:

I am testing with the latest beta and I get another bug with reading http
files. This is a simplified version of what I use in the filter module:

    \unprotect
     \def\AnyFilename#1%
         {\ctxcommand{getreadfilename("any",".","#1")}}

     \def\ProcessFile#1%
       {\edef\InputFile{\AnyFilename{#1}}%
        \InputFile
        \typefile[\InputFile]}
     \protect
     \starttext
     \ProcessFile{http://cim.mcgill.ca/~adityam/index.html}
     \stoptext

With the latest beta, I get the following error:

    ...inimals/texmf-context/tex/context/base/mkiv/file-res.lua:147: bad
    argument #1 to 'find' (string expected, got nil)
    stack traceback:
             [C]: in function 'string.find'
             ...inimals/texmf-context/tex/context/base/mkiv/file-res.lua:147:
    in upvalue 'action1'
             [string "local action = tokens._action..."]:4: in field
    'getreadfilename'
             [ctxlua]:1: in main chunk

Hans' reply:

Deep down we do this:

    implement {
        name      = "getreadfilename",
        actions   = { getreadfilename, context },
        arguments = "3 strings",
    }

This "3 strings" is just a shortcut for { "string", "string", "string" } and I forgot to expand
that when commands.getreadfilename is constructed but it happened ok for the \clf_getreadfilename
command. Actions can also be a direct function (that then gets three strings), here it is
constructed as a function that calls getreadfilename and pipes in into the context command.

Now, in case you wonder what this is about: I create parsers this way, so the arguments of
\clf_getreadfilename are parse from the lua (there are articles about that trickery in the lua
history documents). Apart from being faster (it avoids grabbing and parsing arguments) it also has
the advantage that we have a more convenient interface .. just look for \clf_ and you see code
similar then when we used \ctxlua {commands.foo ... but this time all (even nested tables) are just
scanned from the lua end. No reaccuring interpretation of the \directlua command here.