Add strip=yes option
Earlier, I had implemented a strip=yes option for `t-vim`, where the VIM
script was used to remove leading whitespace.

Now, the default behavor in MkIV is to remove the leading whitespace in
all environments. The behavior is controllable using `strip=no` option.

For now, I have kept the code in t-vim that removes leading whitespace.
This is for two reasons:

* As of now, the `t-vim` code is shared between mkii and mkiv.
* It may be desirably to remove the leading whitespace from external
  files (which are directly processed by vim).