Add support for XML export
When export is enabled, e.g., using

    \setupbackend[export=yes]

then

    \definevimtyping[CODE][syntax=..., alternative=colorscheme]
    \startCODE
    ...
    \stopextCODE

is exported as

    <vimtyping detail="colorscheme">
    ...
    </vimtyping>

And,

    \inlineCODE{...}

is exported as

    <inlinevimtyping detail="colorscheme">...</inlinevimtyping>

Each line is wrapped inside a `<verbatimline>...</verbatimline>` and each
syntax group is wrapped inside `<syntaxgroup
detail="vim....">...</syntaxgroup>`. See the updated documentation for
details.

The module ships with a default CSS file which can be included as

    \setupexport[cssfile=\vimtypingcssfile]

This CSS file tries to mimic the default PDF output. If you want to tweak the
output then it is recommended that you copy the default CSS file and modify it
as appropriate.