Fix leaking color definitions in syntax group.
Closes #37

In the old code, the name of the color just depended on the syntaxgroup
but not on the colorscheme. So, if another colorscheme change the color
of the same syntax group, the name of the color got overridden.

The fix includes the name of the colorscheme in the name of the color.