Remove --clean option
`--clean` is equivalent to `-u NONE -i NONE`. It was added because `-i NONE`
was not working correctly in a previous version of nvim. However, the
order of commands created a bug reported (and diagnosed) by Nicola on
the mailing list. If `--clean` appears *after* `-u`, it overrides the
`-u` option.

For now simply reverted the previous commit.