Fix filter module runtime errors under LMTX Beta
LMTX 2021.08.04 (contextgarden/context-mirror@898d8e1) changes the
signature of the Lua function job.files.run. It now takes a table
containing a filename instead of a string. However, this function
only runs ConTeXt now (not arbitrary programs) so it is of no use
for the filter module.

This commit removes the code that calls job.files.run(...).