Fix #57
Two files with the same name but different extensions got mapped to same
basename. Added the extension to basename as well so that this clash
does not happen.

If the two files with the same name (and extension) are located in
different directories, we will still get a name clash. We can fix this
by using a md5 sum instead of filename, but that makes it harder for a
casual user to understand what is happening.