Fix #59
Typesetting two external files (in the vim module) which lie in different
directories but have the same name resulted in a clash. It was not possible
to use the same external file twice with different start-stop values.
Repored on Github (https://github.com/adityam/filter/issues/59). We now
create an MD5 hash of the full filename, and all options which are passed to
vim. So, if either the filename or any of these options change, a output
filenames are different.

commit 44795d01f7117588b4acefc324c81dd847ab5e2d
Author: Aditya Mahajan <adityam@umich.edu>
Date:   Wed Feb 16 23:26:16 2022 -0500

    Use all options passed to vim to compute the hash

commit ec413fc22472962642da3c84ffea4db6ce7c65c0
Author: Aditya Mahajan <adityam@umich.edu>
Date:   Wed Feb 16 22:32:59 2022 -0500

    Add hash to output file name

    As reported in #59, if two files with the same name in different
    directory are processed, the name of the output filename is the same.
    This patch adds 8 character md5 hash of the full path to the output
    filename.